Getting rid of "begin-link" construct to avoid whitespace in href
The publication tool seems to complain that the editor draft's URL is invalid.
This may be due to the existence of a whitespace before the closing quote in
the "href" attribute of that link.

That whitespace is generated by tidy to accommodate the 80-character limit
constraint. Tidy cuts the link because we use the <!--begin-link--> Anolis
construct and so Tidy rightly thinks it can introduce a carriage-return
after the comment without problem. The thing is Anolis takes whatever
character appears between the comments, carriage-return included.

I switched to a regular <a href="url">url</a> construct to get rid of the
carriage-return for good.