Spec source converted to ReSpec editing tool
On top of the JSON header needed to tell ReSpec how to generate the right
boilerplates, main changes that I introduced are:

1. Anolis uses <span> and <code> for cross-references whereas ReSpec only
understands <a>.

2. All terms used in the specification are now defined in the Terminology
section. Other sections link back to that section. No need to specify the
spec anymore each time one wants to cross-reference a specific term.

3. WebIDL definitions are no longer decorated with <dfn> and <a> since ReSpec
takes care of everything.

4. However, cross-references to interface attributes now need to specify the
interface name in a "for" attribute. For sections that reference or define
more than one attribute of the same interface, I used "dfn-for" or "link-for"
on the common ancestor to be able to stick to simple <a> within the prose.

5. Wrapping <code> is no longer needed for cross-references to interface
definitions as ReSpec automatically adds <code> in the generated prose.
However, note that there is a known bug in ReSpec that makes some of the
references to "startSession" look like code when they should not for the time
being:
 https://github.com/w3c/respec/issues/441

6. Switched to using uppercase MUST, SHOULD, MAY, because ReSpec does not like
it when someone writes the "The key words MAY, MUST, SHOULD are to be
interpreted as described in [RFC2119]" in the Conformance section and creates
a duplicate.

7. I defined terms with singular and plural forms when needed to be able to
reference them either way.

8. To reference the term "presentation", one needs to use the longer form
"presentation context" or "presenting browsing context" (possibly in a "title"
attribute). The term collided with the "Presentation" interface that cannot
be easily renamed.