Adapted Makefile, Travis CI and Echidna configurations for ReSpec version
The Makefile is now only useful to tidy the document. Note that tidy reports
warnings due to the use of proprietary attributes such as "for" in ReSpec. That
is normal. Problem is tidy also seems to exit with code 1 whenever a warning is
reported. I tried all tidy config options that may affect how warnings are
processed but did not find a way to have it exit with code 0.

The Makefile could ignore the error by prefixing the command with "-" but since
that's the only command run by the Makefile (and since we would not want to
ignore real errors), I left things as is for the time being.

Publication to Echidna will only happen on the TR branch as before, but note it
will directly use the "index.html" file. Generating a separate Working Draft is
no longer needed. We can revisit that in the future if we feel that generating
a separate Working Draft different from the Editor's Draft would actually be
useful.