Reorder start steps to create session only when needed
As things stood, the steps to "start a presentation session" required the UA
to create a valid presentation session identifier before creating the task C
to navigate the receiving browser context.

However, the presentation session identifier will only be added to the set of
presentations when C completes. If the user calls "start" again in the
meantime, a UA that would generate the same valid presentation session
identifier for the second presentation session would correctly follow these
steps (that identifier would not yet be in the set of presentations), but this
would create two PresentationSession with the same identifier.

Also, the new PresentationSession S is actually not needed until C completes
successfully.

This commit moves the creation of the presentation session identifier and of
the presentation session itself to the steps that run when C completes
successfully.