Fixed event steps, using "fire" or "fire a simple event" as needed
"Fire a simple event" is now used for all events that inherit from Event.
"Fire an event" is now used for all custom events. The prose is a bit more
verbose in that case to clarify that the event is trusted, does not bubble and
non cancelable.

The bits on the PresentationReceiver interface are somewhat unclear but should
be handled as part of #193, I think.