Split the "close" algorithm into "closing" and "close" procedures
This commit tries to make the closing procedure clearer, in particular the fact
that this is a two-step procedure:
1. the user agent starts closing the underlying connection, signaling the other
end that about its plans.
2. the user agent actually fires the "closed" event when the underlying
connection is closed.

The first step eventually calls the second when the underlying connection is
marked as "closed".

The different possibilities that call the first step are:
- presentation connection could not be established
- close() was called
- a PresentationConnection object S is being discarded

The different possibilities that directly call the second step are:
- the underlying connection has been closed
- unrecoverable error while sending a message
- unrecoverable error while receiving a message
- signal received from the other end that the PresentationConnection object
is to be closed

I extracted the definition of "destination browsing context" from one of the
procedures and moved it to Common Idioms section to be able to reference it from
other procedures.

I also clarified that conformance requirements phrased against a "user agent"
may refer to the controlling browsing context or to the receiving browsing
context depending on the context.