Added enum and event definitions, <var> for variables, fixed references
I added possible definitions for the different presentation connection states.

All references to connection states and event names should now link back to
the definitions of these terms. This has the advantage of clarifying when
"closed" is used to mean the presentation connection state, the reason why
a connection was closed, or the event fired when the connection is closed.

The spec also used a combination of <code>, <em> and <var> to refer to variables
in different places. It should now only use <code> for interface, property,
event names, and code examples; and use <var> to refer to variable and parameter
names.

ReSpec confused the term "presentation" defined in Common idioms as an alias
for "receiving browsing context", and the "presentation" attribute of the
Navigator interface, leading to weird-looking references to "presentation"
that actually targeted the attribute instead of the idiom's definition.

Adding a namespace to the idiom definition is apparently enough to fix the
problem. ReSpec picks up the idiom as a target when the namespace is not given
in particular.