Split "set of presentations" into controlling/receiving sets
I split the "set of presentations" into a "set of controlling presentations" and
a "set of incoming presentations". The former set is a singleton shared across
the browsing contexts running in the controlling user agent (and possibly
published to other user agents). The latter set is per receiving browsing
context.

I also dropped the description of tuples that defined the set of presentations.
Unless I missed something, all the information in these tuples is actually
exposed by the presentation connection object.

The prose makes it clear that all presentation connections share the same URL
and ID in the receiving context.

The reconnect() algorithm was also updated to cover the case where a
PresentationConnection has already been created in the current context for the
same URL and receiving browsing context, and the case where the connection state
of the PresentationConnection instance retrieved from the set of controlling
presentations is "connecting" or "closed".

With these changes, given a controlling browsing context, there should be at
most one presentation connection for a given URL and receiving browsing context.
As such, I don't see any need to parse the set of controlling connections to
fire a "connected" event on those that have the same presentation identifier.
So I simplified that part in the "Establish a presentation connection"
procedure.

This commit intends to fix the following issues:
- #165 conflict description for uniquely identifying a presentation session
- #193 Define a "set of presentations" for the receiving context?
- #229 Need algorithm for establishing a presentation connection when the
previous connection(s) were closed
- #245 "Known connection" in reconnect() may be linked to another browsing
context