Renamed sets of presentations using suggested names.
The "set of controlling presentations" is now called the "set of controlled
presentations", and the "set of incoming presentations" is the "set of
presentation controllers".

Also fixed the wrong claim that the set of controlled presentations cannot
contain two presentations with the same URL and ID. It can contain only one
of such object per controlling browsing context, but there may indeed be
a similar object in other browsing contexts.