Drop use of passive form for event names (issue #256)
Renamed "connected" into "connect", "closed" into "close" and "terminated" into
"terminate" as suggested.

Note I had to introduce a namespace to distinguish the close/terminate methods
and the eponymous events.

This commit also contains a minor fix to references to "connectionavailable":
it is defined once for PresentationConnection and once for
PresentationConnectionList. Unqualified references go the first definition,
but some of them should actually reference the second one. All references are
now explicit as to which definition they target.