Styles adjusted for new official 2016 stylesheet
The custom styles that we had for examples should no longer be needed, and the
ones for IDL declarations cause the "WebIDL" banner to appear truncated.

Also fixed the structure of the Security and privacy considerations section,
as ReSpec needs <section> tags to detect the headings and build the appropriate
Table of Contents.