Improvements to "getting the connectionList attribute" steps (#254)
Getting the connectionList attribute now returns the same
PresentationConnectionList instance, and only resolves the promise when the
set of presentation controllers is non empty (as things stood, the use of a
Promise was useless).

The commit also drops the step that instructed the receiving user agent to start
monitoring incomping presentation connections if not already started. This is
done in the "starting a presentation" algorithm, no need to repeat it here.