Clarified which UA performs the receiving browsing context step
The "start a presentation" algorithm was phrased against the controlling user
agent, but some of the last steps are actually to be performed by the receiving
user agent. Similarly, the "create a receiving browsing context" algorithm
mostly applied to the receiving user agent, but the first step was to be*
performed by the controlling user agent.

This commit ensures that the subject remains the same throughout the steps of
these algorithms.

(The underlying goal is to ensure that algorithms can be linked to one and only
one conformance class, to ease testing)