Fix link to WebIDL "throw" and update ref to WebIDL to target Second Level
The <dfn> tags were incorrectly surrounding "term" instead of "throw".

Also, as alluded to in #295, "NotAllowedError" is only defined in
"WebIDL (Second Level)", and the spec referenced "WebIDL Level 1". I'm not
clear at this stage whether "NotAllowedError" will make it to Level 1, but it
seems better to reference Level 2 in the meantime.