CR Exit criteria
This commit introduces a "CR exit criteria" appendix that details the criteria
that need to be fulfilled for this specification to be published as a Proposed
Recommendation (and thus to exit the Candidate Recommendation status).

The exit criteria are based on similar criteria and wording used in other specs,
e.g.:
- CSS flexbox: http://www.w3.org/TR/css-flexbox-1/#cr-exit-criteria
- HTML5: https://dev.w3.org/html5/decision-policy/public-permissive-exit-criteria.html

Note other specs do not detail what they mean by "independent", "interoperable"
and "implementation" but I think that it's a good idea to be explicit.

On top of the usual "two independent interoperable implementations" criterion,
I added the need to demonstrate that there are implementations of the 1-UA mode
and of the 2-UA mode for receiving user agents.

This commit also updates the Status of This Document (SoTD) section to reference
the exit criteria. Note further edits of the SoTD will be needed prior to
publication as Candidate Recommendation.