Clarified the default presentation request mechanism
This update creates a definition of "default presentation request" in the
"Common Idioms" section, and reformulates the intended behavior in controlling
user agents that support this mechanism.

In particular, the update tries to make it clear that support for this feature
is optional, and moves the check on the sandboxed presentation browsing context
flag to a note since it will be done as part of the "start a presentation"
algorithm.