Clarify where partial Presentation interfaces are to be implemented
To avoid confusing sentences about Presentation attributes that we don't expect
to be present in controlling or receiving user agents, this update makes it
clear which category of user agents is to implement which partial interface
definition.

Also dropped "chrome" from sentences that used "browser chrome", not to
restrict the mechanism to interactions with the browser window.