Switch to automatic cross-references (#486)
* Switch to automatic cross-references

Since the specification was written, authoring tools have made a lot of progress
on cross-references support, and groups have become better at differentiating
between definitions that should remain internal to a given specification
(think "private" in object oriented languages) and those that the group commits
to maintain over time and that can be referenced by other specifications
(think "public").

This update uses the `xref` feature of Respec throughout the spec, and drops
all indirect definitions that were in the Terminology section that can now be
taken care of automatically by Respec. Some of them cannot be because the
underlying specifications are not known to Respec.

Some of the definitions no longer existed (for instance the concept of user
activation was modified in HTML), or specs are no longer the right ones
(reference to Web Storage should now be a reference to HTML).

The exercise also reveals cases where the definitions exist but are not yet
exported. Most of them are linked to the creation of the receiving browsing
context algorithm. This is not surprising since this part cherry-picks
algorithms in different specifications to define an "incognito mode".

In any case, these remaining errors signal the need to exchange with the groups
responsible for these specs to ask them whether they can flag the definitions
we need as "exported".

The definitions that the Presentation API references but that are not exported
by the underlying specs are:

In HTML:
- [parse a URL](https://html.spec.whatwg.org/multipage/urls-and-fetching.html#parse-a-url)
- [list of descendant browsing contexts](https://html.spec.whatwg.org/multipage/browsers.html#list-of-the-descendant-browsing-contexts)
- [creating a new browsing context](https://html.spec.whatwg.org/multipage/browsers.html#creating-a-new-browsing-context)
- [session history](https://html.spec.whatwg.org/multipage/history.html#session-history)
- [allowed to navigate](https://html.spec.whatwg.org/multipage/browsers.html#allowed-to-navigate)
- [navigating to a fragment identifier](https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigate-fragid)

In IndexedDB:
- [databases](https://www.w3.org/TR/IndexedDB/#database)

Note this update also adds the IDL index and the index appendices that Respec
generates automatically.

* Add explicit dfns for terms not in xref database

The cross-ref database only contains terms that specifications choose to export.
The Presentation API needs to reference a few additional terms that are not yet
exported by the relevant specs (mostly HTML). While the group figures out a way
to have these terms exported, this update creates explicit local definitions for
these terms in the terminology section.