Update 301 links and test suite URL
301 links need to be updated per publication rules.

The `wpt.live` domain should be more stable than `w3c-test.org`.

This update also fixes an HTML markup error: `link-for` should have been
`data-link-for` (otherwise it remains as `link-for` in the generated page,
which is invalid HTML). In practice, the attribute was not needed in the
section where it appeared. This update removes it.