Enable auto-publication of the spec to /TR (#502)
This update enables auto-publication of the spec as Candidate Recommendation
Draft to /TR, as agreed by the working group. It leverages the spec-prod action
to do that.

This update also adjusts the workflow for the Editor's Draft itself. The source
spec is now to be found in the `main` branch, and instead of publishing the
source spec to GitHub Pages directly, the spec-prod action will rather deploy
the generated spec. Among other things, this makes it possible to integrate the
spec in Bikeshed's database for cross-referencing purpose.

The spec-prod action also takes care of running ReSpec on pull requests to
validate changes from an editorial perspective.