Update cross references to HTML. (#514)
* WIP

* Fix additional respec errors.

* Replace "descendant browsing contexts", restore link to "unload a document"

The notion of "list of descendant browsing contexts" also needed to be replaced
by "descendant navigables". This was not reported by ReSpec because the previous
term was not exported by HTML so it was defined locally in the Terminology
section. Good news is "descendant navigables" is exported, so we can link to it
directly.

It seems a pity to lose the link to "unload a document" just because HTML no
longer export the term. This update adds it back as a local definition that
links to the (private for now) definition in HTML. We'll need to check that
list of terms at some point and exchange with HTML editors to see whether they
could export them, think we're doing something wrong, or think we should
continue with the current setting.

Co-authored-by: Francois Daoust <fd@tidoust.net>